Move `Handle<T>` to a separate file
