+++ src/asset/core/index.js
@@ -1,4 +1,5 @@
 export * from './asset.js'
 export * from './channel.js'
+export * from './handle.js'
 export * from './exporter.js'
 export * from './importer.js'
